Restructure '5-share-versions-with-build-src'
It now works with pre-compiled script plugin.

It is important that the build that provides a plugin for another
build is included before that build. In this case, 'dependency-versions'
needs to go first.

Also when reading a file eagerly, we need to be careful that the code
does not fail during script compilation where we configure an artificial
project. That's why there is a File.exists() check now.